feat: separated the content of the NativeShortcutGuide into smaller components
