+++ src/assets/data/GuideData.ts
@@ -1,14 +1,21 @@
-interface Shortcut {
+/**
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, version 3 of the License.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
+ * You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
+ * Other licensing options may be available, please reach out to data-viz@feedzai.com for more information.
+ */
+
+export interface Shortcut {
 	keys: string;
 	description: string;
 }
 
-interface Guide {
+export interface Section {
 	title: string;
 	shortcuts: Shortcut[];
 }
 
-export const GUIDE_DATA: Guide[] = [
+export const GUIDE_DATA: Section[] = [
 	{
 		title: "Access the guide",
 		shortcuts: [
