+++ src/components/shortcut_guide/components/NativeGuideDescription.tsx
@@ -5,5 +5,17 @@
  * Other licensing options may be available, please reach out to data-viz@feedzai.com for more information.
  */
 
-export { default as AutoVizuA11y } from "./AutoVizuA11y";
-export { default as NativeShortcutGuide } from "./components/shortcut_guide/components/NativeShortcutGuide";
+import * as constants from "../../../constants";
+
+export const ShortcutGuideDescription = () => (
+	<p
+		className={constants.SHORTCUTGUIDE_CLASSES.shortcutGuideDescription}
+		id={constants.SHORTCUTGUIDE_ID.shortcutGuideDescription}
+	>
+		AutoVizually shortcut guide. AutoVizually lets you navigate between charts and underlying data
+		elements using just the keyboard. When focused on a chart, a description regarding the data will
+		be provided — you might receive a notification indicating that the chart description was
+		produced by an AI model. For JAWS and NVDA users, it is recommended to turn Focus mode before
+		navigating the data using the arrow keys.
+	</p>
+);
