+++ src/components/shortcut_guide/components/NativeGuideHeader.tsx
@@ -5,5 +5,23 @@
  * Other licensing options may be available, please reach out to data-viz@feedzai.com for more information.
  */
 
-export { default as AutoVizuA11y } from "./AutoVizuA11y";
-export { default as NativeShortcutGuide } from "./components/shortcut_guide/components/NativeShortcutGuide";
+import * as constants from "../../../constants";
+
+export const ShortcutGuideHeader = ({ onClose }: { onClose: () => void }) => (
+	<div className={constants.SHORTCUTGUIDE_CLASSES.shortcutGuideHeader}>
+		<h2
+			className={constants.SHORTCUTGUIDE_CLASSES.shortcutGuideTitle}
+			id={constants.SHORTCUTGUIDE_ID.shortcutGuideTitle}
+		>
+			Shortcut Guide
+		</h2>
+		<p className={constants.SHORTCUTGUIDE_CLASSES.shortcutGuideActionLabel}>? or Esc</p>
+		<button
+			className={constants.SHORTCUTGUIDE_CLASSES.shortcutGuideButtonClose}
+			aria-label="Close shortcut guide"
+			onClick={onClose}
+		>
+			&times;
+		</button>
+	</div>
+);
